requirements.txt
fastapi
uvicorn
sqlalchemy
python-jose[cryptography]  # Для JWT
passlib                    # Хеширование паролей
pydantic[email]
bcrypt
jinja2
Pillow
numpy
opencv-python-headless